% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yzmiscf_functions.R
\name{cleanme}
\alias{cleanme}
\title{A Miscellaneous Function}
\usage{
cleanme(raw_data, clean_output)
}
\arguments{
\item{raw_data}{##raw Ct values from PCR, duplicates or triplicates}
}
\description{
This function allows you to take PCR raw Ct values of technical replicates,
de-multiplicates each sample, and returns with a unique average Ct value for each biological sample.
}
\examples{
cleanme()
}
\keyword{data}
\keyword{wrangling}
