% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanme.R
\name{cleanpcr}
\alias{cleanpcr}
\title{001. de-multiplex technical replicates}
\usage{
cleanpcr(raw_data)
}
\arguments{
\item{raw_data}{A tbl, preferably a .csv file with a proper header. Set column 1 as "id",
column 2 as "variable", the rest columns can be raw Ct values for target genes.}
}
\description{
This function allows me to take PCR raw Ct values of technical replicates,
    and returns with a unique average Ct value for each biological sample.
}
\details{
1st column of the raw data needs to be integer id.
}
\keyword{data}
\keyword{wrangling}
