% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yzmiscf_functions.R
\name{dogs_over_cats}
\alias{dogs_over_cats}
\title{A Dog Function}
\usage{
dogs_over_cats(agree = TRUE)
}
\arguments{
\item{agree}{Do you agree dogs are the best pet? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior furry animal.
}
\examples{
dogs_over_cats()
}
\keyword{dogs}
