% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yzmiscf_functions.R
\name{expfc}
\alias{expfc}
\title{002. Gene expression fold changes.}
\usage{
expfc(.clean.df, dct_cont)
}
\arguments{
\item{cleaned_PCR_Ct_values}{A tbl or a df, can be result from cleanme function.}
}
\description{
This function allows me to take cleaned PCR Ct values of target genes,
    and returns with a fold change relevant to the reference gene for each target gene.
}
\details{
blank.
}
\keyword{Ct}
\keyword{delta}
