% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yzmiscf-package.R
\docType{package}
\name{yzmiscf-package}
\alias{yzmiscf}
\alias{yzmiscf-package}
\title{yzmiscf: a package of miscellaneous functions}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

yzmiscf keeps track of a list of handy miscellaneous functions I use constantly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zhangyang2017/yzmiscf}
  \item Report bugs at \url{https://github.com/zhangyang2017/yzmiscf/issues}
}

}
\author{
\strong{Maintainer}: Yang Zhang \email{yang.yz.zhang@gmail.com}

}
\keyword{internal}
